# internal/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "JobPulse API",
        "description": "Job aggregation platform — search real listings from multiple sources, manage user accounts, and analyze market trends.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local dev"
        }
    ],
    "tags": [
        {
            "name": "Health",
            "description": "Service health"
        },
        {
            "name": "Jobs",
            "description": "Job listings"
        },
        {
            "name": "Companies",
            "description": "Company profiles"
        },
        {
            "name": "Analytics",
            "description": "Market insights"
        },
        {
            "name": "Auth",
            "description": "Authentication"
        },
        {
            "name": "User",
            "description": "Protected user routes"
        }
    ],
    "paths": {
        "/health": {
            "get": {
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "List jobs with filters & pagination",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search query (title, skills, company)"
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Filter by location"
                    },
                    {
                        "name": "experience",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "entry",
                                "mid",
                                "senior",
                                "lead"
                            ]
                        },
                        "description": "Experience level"
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "linkedin",
                                "indeed",
                                "jsearch",
                                "adzuna"
                            ]
                        },
                        "description": "Job source"
                    },
                    {
                        "name": "remote",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true"
                            ]
                        },
                        "description": "Remote only"
                    },
                    {
                        "name": "salary_min",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Minimum salary"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "date",
                                "salary",
                                "relevance"
                            ]
                        },
                        "description": "Sort order"
                    },
                    {
                        "name": "refresh",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true"
                            ]
                        },
                        "description": "Trigger live API fetch (requires API keys)"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated job listings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/JobSummary"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "List companies",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search by name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Company"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/companies/{slug}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get company with jobs",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details with jobs"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/filters": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get available filter options",
                "responses": {
                    "200": {
                        "description": "Filter options (locations, sources, experience levels)"
                    }
                }
            }
        },
        "/api/v1/analytics/skills": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Top skills by frequency",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill counts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SkillCount"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/analytics/summary": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Analytics overview",
                "responses": {
                    "200": {
                        "description": "Summary stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalyticsSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password",
                                    "name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "secret123"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Email already exists"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get current user profile",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/me/saved-jobs": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List saved jobs",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saved job list"
                    }
                }
            }
        },
        "/api/v1/me/saved-jobs/{id}": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Save a job",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job saved"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Unsave a job",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job unsaved"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Job": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "companySlug": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "salaryMin": {
                        "type": "integer",
                        "nullable": true
                    },
                    "salaryMax": {
                        "type": "integer",
                        "nullable": true
                    },
                    "salaryCurrency": {
                        "type": "string"
                    },
                    "postedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "source": {
                        "type": "string"
                    },
                    "sourceUrl": {
                        "type": "string"
                    },
                    "skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "isRemote": {
                        "type": "boolean"
                    },
                    "employmentType": {
                        "type": "string"
                    },
                    "experienceLevel": {
                        "type": "string"
                    }
                }
            },
            "JobSummary": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "salaryMin": {
                        "type": "integer",
                        "nullable": true
                    },
                    "salaryMax": {
                        "type": "integer",
                        "nullable": true
                    },
                    "isRemote": {
                        "type": "boolean"
                    },
                    "postedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "source": {
                        "type": "string"
                    }
                }
            },
            "Company": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "industry": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    },
                    "jobCount": {
                        "type": "integer"
                    }
                }
            },
            "SkillCount": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "count": {
                        "type": "integer"
                    }
                }
            },
            "AnalyticsSummary": {
                "type": "object",
                "properties": {
                    "totalJobs": {
                        "type": "integer"
                    },
                    "totalCompanies": {
                        "type": "integer"
                    },
                    "jobsToday": {
                        "type": "integer"
                    },
                    "jobsThisWeek": {
                        "type": "integer"
                    },
                    "averageSalary": {
                        "type": "integer"
                    },
                    "remoteJobsCount": {
                        "type": "integer"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                }
            },
            "Pagination": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}